shell Command
Add command that starts a new agent and tries to start the users shell
with some helpers around making sure the environment is available with
every command and ensuring the agent is killed when the shell is closed.
Currently only supports bash but should be extended to support other
shells.